obedient_beast/requirements.txt
# Obedient Beast - Python Dependencies
httpx>=0.27.0
anthropic>=0.25.0
openai>=1.30.0
python-dotenv>=1.0.0
flask>=3.0.0
